requirements.txt
Django
djangorestframework
psycopg2-binary
django-filter
Pillow
django-decouple
drf_yasg
